Allow values up to 12 bits

+++ mod_mixmodssd/mod_mixmodssd.py
@@ -1,7 +1,7 @@
 
 def generate_c_code(icosoc_h, icosoc_c, mod):
     icosoc_h.append("""
-static inline void icosoc_%s_set(uint8_t x)
+static inline void icosoc_%s_set(uint16_t x)
 {
     *(uint32_t*)(0x20000000 + %s * 0x10000) = x;
 }
